Split send into multiple mods
